Wed, Aug 2, 2023, 10:36 PM +06:00

+++ codeforces/new_commers_training/sheet1/c.c
@@ -2,10 +2,10 @@
 
 int main(void)
 {
-    int a, b;
-    scanf("%d %d", &a, &b);
-    printf("%d + %d = %d\n", a, b, a + b);
-    printf("%d * %d = %d\n", a, b, a * b);
-    printf("%d - %d = %d\n", a, b, a - b);
+    long long int a, b;
+    scanf("%lld %lld", &a, &b);
+    printf("%lld + %lld = %lld\n", a, b, a + b);
+    printf("%lld * %lld = %lld\n", a, b, a * b);
+    printf("%lld - %lld = %lld\n", a, b, a - b);
     return 0;
 }
